CORE-76: storage: refactor API
